+++ lib/ui/events2/Event2AttendanceTakerPanel.dart
@@ -5,6 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_barcode_scanner/flutter_barcode_scanner.dart';
 import 'package:flutter_widget_from_html/flutter_widget_from_html.dart';
 import 'package:illinois/service/Analytics.dart';
+import 'package:illinois/service/Config.dart';
 import 'package:illinois/ui/events2/Event2CreatePanel.dart';
 import 'package:illinois/ui/events2/Event2Widgets.dart';
 import 'package:illinois/ui/widgets/HeaderBar.dart';
@@ -297,21 +298,24 @@ class _Event2AttendanceTakerWidgetState extends State<Event2AttendanceTakerWidge
   Widget _buildUploadAttendeesDescription() {
     TextStyle? mainStyle = Styles().textStyles?.getTextStyle('widget.item.small.thin.italic');
     final Color defaultStyleColor = Colors.red;
-    final String adminAppUrl = 'go.illinois.edu/ILappAdmin'; //TBD: DD - move it to config
-    final String adminAppUrlMacro = '{{admin_app_url}}';
-    String contentHtml = Localization().getStringEx('panel.event2.detail.attendance.attendees.description', "Looking for a way to upload an attendee list or download your current attendees? Share the link or visit <a href='{{admin_app_url}}'>{{admin_app_url}}</a>.");
-    contentHtml = contentHtml.replaceAll(adminAppUrlMacro, adminAppUrl);
-    return Padding(padding: EdgeInsets.only(top: 12), child:
-      Row(crossAxisAlignment: CrossAxisAlignment.start, children: [
-        Styles().images?.getImage('info') ?? Container(),
-        Expanded(child:
-          Padding(padding: EdgeInsets.only(left: 6), child:
-            HtmlWidget(contentHtml, onTapUrl: _onTapHtmlLink, textStyle: mainStyle,
-              customStylesBuilder: (element) => (element.localName == "a") ? { "color": ColorUtils.toHex(mainStyle?.color ?? defaultStyleColor), "text-decoration-color": ColorUtils.toHex(Styles().colors?.fillColorSecondary ?? defaultStyleColor)} : null,
-            )
+    final String? eventAttendanceUrl = Config().eventAttendanceUrl;
+    final String eventAttendanceUrlMacro = '{{event_attendance_url}}';
+    String contentHtml = Localization().getStringEx('panel.event2.detail.attendance.attendees.description',
+        "Looking for a way to upload an attendee list or download your current attendees? Share the link or visit <a href='$eventAttendanceUrlMacro'>$eventAttendanceUrlMacro</a>.");
+    contentHtml = contentHtml.replaceAll(eventAttendanceUrlMacro, eventAttendanceUrl ?? '');
+    return Visibility(visible: StringUtils.isNotEmpty(eventAttendanceUrl), child:
+      Padding(padding: EdgeInsets.only(top: 12), child:
+        Row(crossAxisAlignment: CrossAxisAlignment.start, children: [
+          Styles().images?.getImage('info') ?? Container(),
+          Expanded(child:
+            Padding(padding: EdgeInsets.only(left: 6), child:
+              HtmlWidget(contentHtml, onTapUrl: _onTapHtmlLink, textStyle: mainStyle,
+                customStylesBuilder: (element) => (element.localName == "a") ? { "color": ColorUtils.toHex(mainStyle?.color ?? defaultStyleColor), "text-decoration-color": ColorUtils.toHex(Styles().colors?.fillColorSecondary ?? defaultStyleColor)} : null,
+              )
+            ),
           ),
-        ),
-      ])
+        ])
+      ),
     );
   }